# gatewayservice/openapi.yaml
openapi: 3.0.0
info:
  title: Gatewayservice API
  description: Gateway OpenAPI specification.
  version: 0.2.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://98.66.177.253:8000
    description: Production server
paths:
  /login:
    post:
      summary: Log in to the system.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User ID.
                  example: student
                password:
                  type: string
                  description: User password.
                  example: pass
      responses:
        "200":
          description: Login successful. Returns user token, username, and creation date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: User token.
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2N2UzY2ZiZGY1MzJmZmMwYTk1MzNjMmUiLCJ1c2VybmFtZSI6IlRlc3QiLCJpYXQiOjE3NDM0MzIwODUsImV4cCI6MTc0MzQzNTY4NX0.2O6lbAOkuEdJBaZNP7YmfAGgrL6gGt9E6dEa1Lrb7R4
                  username:
                    type: string
                    description: Username.
                    example: student
                  createdAt:
                    type: string
                    description: Creation date.
                    example: "2025-03-26T09:58:21.422Z"
        "400":
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error for the localized template telling some field is missing.
                    example: "logIn.errors.missingFields"
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: A messaage with the localization for the invalid credentials message.
                    example: "logIn.errors.invalidCredentials"
        '500':
          description: Some error ocurred in the server when processing the request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /adduser:
    post:
      summary: Add a new user to the database.
      operationId: addUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User ID.
                  example: student
                password:
                  type: string
                  description: User password.
                  example: pass
      responses:
        "200":
          description: User added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: User token.
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2N2VhYTkyM2FlZTYxNThkZTJjOGEyMDciLCJpYXQiOjE3NDM0MzE5NzEsImV4cCI6MTc0MzQzNTU3MX0.7lC7CSStyWlxOYxwgU7N1iPodiKHTD9Jq7e09Wa-VFc
                  username:
                    type: string
                    description: User ID.
                    example: student
                  createdAt:
                    type: string
                    description: Creation date.
                    example: "2025-03-31T14:39:31.025Z"
        "500":
          description: Failed to add user, there was an error in the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: getaddrinfo ENOTFOUND mongodb
  /default-images/{imageName}:
    get:
      summary: Retrieve a default profile image.
      operationId: defaultImage
      parameters:
        - name: imageName
          in: path
          required: true
          description: The name of the image to retrieve.
          schema:
            type: string
            example: image_1.png
      responses:
        "200":
          description: Image retrieved correctly.
          content:
            image/*:
              schema:
                type: string
                format: binary
                description: The image data.
        "404":
          description: Image not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Request failed with status code 404
        "500":
          description: Failed to retrieve the image due to a server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /users/{username}:
    patch:
      summary: Update the user information.
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the account to update.
          schema:
            type: string
            example: currentUsername
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The new username, if wanted to be modified.
                  example: newUsername
                password:
                  type: string
                  description: The new password, if wanted to be modified.
                  example: newPassword123!
                passwordRepeat:
                  type: string
                  description: Confirmation of the new password.
                  example: newPassword123!
      responses:
        '200':
          description: User updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful.
                    example: true
                  currentUsername:
                    type: string
                    description: The accout username after the operations.
                    example: newUsername
                  token:
                    type: string
                    description: New user token after the operations.
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2N2VhYTkyM2FlZTYxNThkZTJjOGEyMDciLCJpYXQiOjE3NDM0MzE5NzEsImV4cCI6MTc0MzQzNTU3MX0.7lC7CSStyWlxOYxwgU7N1iPodiKHTD9Jq7e09Wa-VFc
                  message:
                    type: string
                    description: Message describing the result.
                    example: User updated successfully
        '400':
          description: Invalid request parameters. (Passwords don't match, invalid username/password, etc...)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Invalid input data.
        '401':
          description: User token is missing or invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Authorization header missing
        '403':
          description: Invalid user modifying the data. (username is not the same as the one in the token)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: You can only update your own account.
        '500':
          description: Failed to update user due to a server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /users/{username}/image:
    get:
      summary: Retrieve the profile image of some user.
      operationId: userImage
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the user to recover their image.
          schema:
            type: string
            example: student
      responses:
        "200":
          description: Image retrieved correctly.
          content:
            image/*:
              schema:
                type: string
                format: binary
                description: The image data.
        "404":
          description: User is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: User not found
        "500":
          description: Failed to retrieve the image due to a server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /users/{username}/default-image:
    post:
      summary: Updates the image of some user to a default one.
      operationId: defaultUserImage
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the user to recover their image.
          schema:
            type: string
            example: student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  description: The name of the default image to set.
                  example: image_13.png
      responses:
        "200":
          description: Image updated correctly.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful.
                    example: true
                  message:
                    type: string
                    description: Message describing the result.
                    example: Image updated successfully
                  imagePath:
                    type: string
                    description: Path to the updated image.
                    example: /images/default/image_13.png
        "400":
          description: Default image is missing in the body.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: No default image provided
        "404":
          description: User is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: User not found
        "500":
          description: Failed to upload the image due to a server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /users/{username}/custom-image:
    post:
      summary: Updates the image of some user, uploading a new one.
      operationId: customUserImage
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the user to recover their image.
          schema:
            type: string
            example: student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: The image file to upload.
      responses:
        "200":
          description: Image updated correctly.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful.
                    example: true
                  message:
                    type: string
                    description: Message describing the result.
                    example: Image updated successfully
                  imagePath:
                    type: string
                    description: Path to the updated image.
                    example: /images/default/image_13.png
        "400":
          description: File is missing or is not an image.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Uploaded file is not an image
        "404":
          description: User is not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: User not found
        "500":
          description: Failed to upload the image due to a server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error information.
                    example: Internal Server Error
  /askllm:
    post:
      summary: Send a question to the LLM.
      operationId: askLLM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gameQuestion:
                  type: string
                  description: Question being showed in the game.
                  example: Which of the following flags belongs to Austria?
                userQuestion:
                  type: string
                  description: Question performed by the user.
                  example: Does it contain the red color?
      responses:
        "200":
          description: The LLM responded successfully, answering the question.
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: LLM answer.
                    example: The color of the flag is a key indicator. It starts with a simple yet distinctive design that many people recognize.
        "400":
          description: Some required data is missing (gameQuestion or userQuestion).
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    description: List of missing fields.
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          description: Type of error.
                          example: field
                        msg:
                          type: string
                          description: Error message.
                          example: Question is required
                        path:
                          type: string
                          description: Field that caused the error.
                          example: gameQuestion
                        location:
                          type: string
                          description: Location of the error in the request.
                          example: body
              example:
                errors:
                  - type: field
                    msg: The game question is required
                    path: gameQuestion
                    location: body
                  - type: field
                    msg: The user question is required
                    path: userQuestion
                    location: body
        "500":
          description: API key is missing or some error ocurred in the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    description: Error content.
                    example: API key is missing.
  /simplellm:
    post:
      summary: Send a simple message to the LLM.
      operationId: simpleLLM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: Message to be sent to the LLM.
                  example: Greet the user called john.
      responses:
        "200":
          description: The LLM responded successfully, with a message.
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: LLM answer.
                    example: Hello john, welcome to the game!
        "400":
          description: Some required data is missing (message).
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    description: List of missing fields.
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          description: Type of error.
                          example: field
                        msg:
                          type: string
                          description: Error message.
                          example: Question is required
                        path:
                          type: string
                          description: Field that caused the error.
                          example: gameQuestion
                        location:
                          type: string
                          description: Location of the error in the request.
                          example: body
              example:
                errors:
                  - type: field
                    msg: The message is required
                    path: message
                    location: body
        "500":
          description: API key is missing or some error ocurred in the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    description: Error content.
                    example: API key is missing.
  /answer:
    post:
      summary: Checks if the answer for some question is correct.
      operationId: answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                questionId:
                  type: string
                  description: The ID of the question.
                  example: jwhwz7icbvitz82b58d2
                answer:
                  type: string
                  description: The image the user selected.
                  example: http://commons.wikimedia.org/wiki/Special:FilePath/Flag%20of%20Costa%20Rica.svg
      responses:
        "200":
          description: The answer was checked correctly.
          content:
            application/json:
              schema:
                type: object
                properties:
                  correct:
                    type: boolean
                    description: Whether the answer was correct or not.
                    example: true
        "500":
          description: The answer could not be validated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message.
                    example: Failed to validate answer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT